adjusts

+++ routes/api.php
@@ -1,8 +1,8 @@
 <?php
 
 use App\Http\Controllers\Auth\LoginController;
-use App\Http\Controllers\SystemProfileController;
-use App\Http\Controllers\UserController;
+use App\Http\Controllers\Admin\SystemProfileController;
+use App\Http\Controllers\Admin\UserController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 
@@ -26,11 +26,11 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
 Route::group(['middleware' => 'auth:sanctum'], function () {
 
     // users
-    Route::get('users', [UserController::class, 'getAll'])->name('admin.users');
-    Route::post('user', [UserController::class, 'save'])->name('admin.users.save');
-    Route::get('user/{id}', [UserController::class, 'get'])->name('admin.users.get');
-    Route::put('user/{id}', [UserController::class, 'update'])->name('admin.users.update');
-    Route::delete('user/{id}', [UserController::class, 'delete'])->name('admin.users.delete');
+    Route::get('admin/users', [UserController::class, 'getAll'])->name('admin.users');
+    Route::post('admin/user', [UserController::class, 'save'])->name('admin.users.save');
+    Route::get('admin/user/{id}', [UserController::class, 'get'])->name('admin.users.get');
+    Route::put('admin/user/{id}', [UserController::class, 'update'])->name('admin.users.update');
+    Route::delete('admin/user/{id}', [UserController::class, 'delete'])->name('admin.users.delete');
 
     // system profile
     Route::get('system/profile', [SystemProfileController::class, 'getAll'])->name('admin.system.profile');
